Porting System.Drawing to SixLabors.ImageSharp

+++ src/CKLunchBot.Core/Utils/CKLunchBotColors.cs
@@ -2,7 +2,7 @@ namespace CKLunchBot.Core.Utils
 {
     public class CKLunchBotColors
     {
-        public static (int, int, int) White { get; } = (238, 238, 238);
-        public static (int, int, int) Black { get; } = (45, 45, 48);
+        public static (byte, byte, byte) White { get; } = (238, 238, 238);
+        public static (byte, byte, byte) Black { get; } = (45, 45, 48);
     }
 }
